테스트시 log안찍히게 변경, server-test 추가, server-test에 맞게 gulp task 추가, 테스트 순서를 위해 connection.spec 파일 위치 변경
#1
