+++ server/.gitignore
@@ -49,4 +49,3 @@ node_modules
 
 #dist
 dist
-
