+++ server/src/server-dev.ts
@@ -8,17 +8,5 @@ dotenv.config({
     path: '../.env'
 });
 
-let _resolve: any;
-
-const readyPromise = new Promise(resolve => {
-    _resolve = resolve;
-});
-
-const appServer = new AppServer();
-appServer.listen().then(() => {
-    _resolve();
-});
-
-/* 테스트를 위해서 아래와 같이 처리 한다. */
-export const ready = readyPromise;
-export const close = () => appServer.wss.close();
+/* 서버 구동 */
+new AppServer();
